+++ tests/test_application.py
@@ -5,6 +5,7 @@ import asyncio
 import pytest
 
 from henson.base import Application
+from henson.exceptions import Abort
 
 
 @pytest.mark.asyncio
@@ -49,6 +50,32 @@ def test_consume(event_loop, test_consumer, cancelled_future):
     # The size of the queue won't ever be larger than 1 because of the
     # maxsize argument.
     assert queue.qsize() == 1
+
+
+def test_consumer_aborts(event_loop):
+    """Test that the application stops after the consumer aborts."""
+    consumer_called = False
+    callback_called = False
+
+    class Consumer:
+        @asyncio.coroutine
+        def read(self):
+            nonlocal consumer_called
+            consumer_called = True
+            raise Abort('reason', 'message')
+
+    @asyncio.coroutine
+    def callback(app, message):
+        nonlocal callback_called
+        callback_called = True
+        while True:
+            yield from asyncio.sleep(0)
+
+    app = Application('testing', consumer=Consumer(), callback=callback)
+    app.run_forever(loop=event_loop)
+
+    assert consumer_called
+    assert not callback_called
 
 
 def test_consumer_exception(event_loop):
@@ -170,6 +197,23 @@ def test_postprocess_results(original, expected):
 
     assert callback1_called
     assert callback2_called
+
+
+def test_process_exception_stops_application(event_loop, test_consumer):
+    """Test that the application stops after a processing exception."""
+    @asyncio.coroutine
+    def callback(app, message):
+        return [{}]
+
+    app = Application('testing', consumer=test_consumer, callback=callback)
+
+    @app.result_postprocessor
+    @asyncio.coroutine
+    def postprocess(app, message):
+        raise Exception()
+
+    with pytest.raises(Exception):
+        app.run_forever(loop=event_loop)
 
 
 @pytest.mark.parametrize('postprocess', (None, '', False, 10, sum))